- added an IntArrayStack and an ArrayBucketQueue that can be used with the BucketQueueAStar (but they are equally fast as the ListBucketQueue) - removed some shorts

+++ src/jsettlers/algorithms/SimpleDijkstraTester.java
@@ -16,7 +16,7 @@ public class SimpleDijkstraTester {
 	public static void main(String args[]) {
 		IDijkstraPathMap map = new IDijkstraPathMap() {
 			@Override
-			public boolean fitsSearchType(short x, short y, ESearchType type, IPathCalculateable requester) {
+			public boolean fitsSearchType(int x, int y, ESearchType type, IPathCalculateable requester) {
 				if (x == 120 && y == 100)
 					return true;
 				if (x == 110 && y == 110)
@@ -28,7 +28,7 @@ public class SimpleDijkstraTester {
 			}
 
 			@Override
-			public void setDijkstraSearched(short x, short y) {
+			public void setDijkstraSearched(int x, int y) {
 			}
 		};
 		DummyEmptyAStarMap aStarMap = new DummyEmptyAStarMap(WIDTH, HEIGHT);
